+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### Compatible changes
 - Add `preload_association` at instance level. Example: `user.preload_associations(:posts)`
+- Add support for polymorphic associations in `traverse_association`.
 
 ## 1.0.0